refactor(go): split legacy CLI entrypoint files
Move legacy CLI orchestration, import/export helpers, SQL command handling, and display helpers into focused Go files while keeping behavior unchanged.
